// frontend/src/backend.d.ts
export class ExternalBlob {
    getBytes(): Promise<Uint8Array<ArrayBuffer>>;
    getDirectURL(): string;
    static fromURL(url: string): ExternalBlob;
    static fromBytes(blob: Uint8Array<ArrayBuffer>): ExternalBlob;
    withUploadProgress(onProgress: (percentage: number) => void): ExternalBlob;
}
export interface Payslip {
    month: string;
    payslipId: bigint;
    year: string;
    payslipDocument: ExternalBlob;
    employeeId: string;
}
export interface Letter {
    created: Time;
    createdBy: Principal;
    letterType: string;
    employeeId: string;
    document: ExternalBlob;
    letterId: bigint;
}
export type Time = bigint;
export interface BulkCreateStatus {
    employeeId: string;
    message: string;
    success: boolean;
}
export interface Contact {
    relationship: string;
    name: string;
    phone: string;
}
export interface Address {
    street: string;
    country: string;
    city: string;
    zipCode: string;
    state: string;
}
export interface LeaveRequest {
    status: string;
    requestId: bigint;
    endDate: Time;
    employeeId: string;
    leaveType: string;
    startDate: Time;
}
export interface EmployeeRecord {
    internalId: bigint;
    status: string;
    leaveBalance: LeaveBalance;
    personalEmail?: string;
    designation: string;
    role: EmployeeRole;
    joiningDate: Time;
    businessEmail: string;
    employmentType: string;
    employeeId: string;
    department: string;
    phoneNumber: string;
}
export interface BulkEmployeeRecord {
    internalId: bigint;
    status: string;
    sickLeave: bigint;
    annualLeave: bigint;
    personalEmail?: string;
    role: EmployeeRole;
    joiningDate: Time;
    businessEmail: string;
    unpaidLeave: bigint;
    casualLeave: bigint;
    employeeId: string;
    department: string;
    phoneNumber: string;
}
export interface EmployeeProfile {
    internalId: bigint;
    status: string;
    personalEmail?: string;
    role: EmployeeRole;
    joiningDate: Time;
    businessEmail: string;
    department: string;
    phoneNumber: string;
}
export interface EmployeeDetails {
    designation: string;
    joiningDate: Time;
    businessEmail: string;
    emergencyContacts: Array<Contact>;
    employmentType: string;
    address: Address;
    department: string;
    phoneNumber: string;
}
export interface LeaveBalance {
    sick: bigint;
    annual: bigint;
    unpaid: bigint;
    casual: bigint;
}
export interface UserProfile {
    name: string;
    role: EmployeeRole;
    employeeId?: string;
}
export enum EmployeeRole {
    hrManager = "hrManager",
    manager = "manager",
    superAdmin = "superAdmin",
    employee = "employee",
    hrStaff = "hrStaff"
}